chore: remove old import tree
